requirements.txt
pydantic>=2.0.0
requests>=2.28.0
pytest>=7.0.0
pytest-html>=3.0.0
allure-pytest>=2.13.0
allure-python-commons>=2.13.0
